test: migrate unique iOS plugin tests into RunnerTests
The SPM test target at ios/flureadium/Tests/flureadiumTests could not run (it needs the Flutter framework, and Readium declares iOS-only support so SPM can't resolve a macOS host). Move its genuinely unique coverage into the runnable RunnerTests target and drop the dead SPM target.

New files (no equivalent existed):
- ScrollModeNavigationTests: strippedHref, chapterLink before/after, isBackwardNavigation.
- EventStreamHandlerTests: EventStreamHandler listen/cancel/dispose lifecycle.

Folded into existing RunnerTests classes:
- FlutterTTSNavigator initNavigator() throws for an unsupported publication.
- RunnerTests: ttsCanSpeak returns false for an unspeakable publication.
- FlutterNavigationConfig type-strict decoding (strings are not coerced to Bool).
- FlutterPdfPreferences invalid-value decoding.
- EdgeTapInterceptView property behavior (threshold get/set, callback invoke/clear).

Dropped as duplicate or stale (already covered by RunnerTests):
- EdgeTap hitTest expectations (superseded by the current subview-based version).
- ReadiumExtensionsMappingTests and the PDF/config duplicates.
- The superseded TTS navigator tests.

Register the two new files in project.pbxproj and remove the .testTarget stanza from Package.swift (the manifest is kept).
