+++ flureadium/ios/flureadium/Package.swift
@@ -40,13 +40,6 @@ let package = Package(
                 // the following instructions to add them:
                 // https://developer.apple.com/documentation/xcode/bundling-resources-with-a-swift-package
             ]
-        ),
-        .testTarget(
-            name: "flureadiumTests",
-            dependencies: [
-                "flureadium",
-                .product(name: "ReadiumNavigator", package: "swift-toolkit"),
-            ]
         )
     ]
 )
